pyproject.toml
[project]
name = "pdf-semantic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.7",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "maturin>=1.13.1",
    "openai>=2.32.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "supabase>=2.28.3",
    "torch>=2.11.0",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["maturin>=1.13.1"]
build-backend = "maturin"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"

[tool.maturin]
manifest-path = "rust_parser/Cargo.toml"
module-name = "rust_parser"
